Added organization page & Added modification of organization

+++ src/service/OrganizationService.ts
@@ -5,23 +5,32 @@ import $api from "../config";
 export class CompanyService {
 
     async getAll(pageable: Pageable): Promise<Page<Organization>> {
-        return await $api.get(`/api/v1/admin/organizations?page=${pageable.page}&size=${pageable.size}`);
+        let response = await $api.get<Page<Organization>>(`/api/v1/admin/organizations?page=${pageable.page}&size=${pageable.size}`);
+        return response.data;
+    }
+
+    async getOne(orgId: number): Promise<Organization> {
+        let response = await $api.get<Organization>(`/api/v1/admin/organizations/${orgId}`);
+        return response.data;
     }
 
     async create(org: Organization): Promise<Organization> {
-        return await $api.post(`/api/v1/admin/organizations`, org);
+        let response = await $api.post<Organization>(`/api/v1/admin/organizations`, org);
+        return response.data;
     }
 
     async modify(orgId: number, org: Organization): Promise<Organization> {
-        return await $api.put(`/api/v1/admin/organizations/${orgId}`, org);
+        let response = await $api.put<Organization>(`/api/v1/admin/organizations/${orgId}`, org);
+        return response.data;
     }
 
     async approve(orgId: number): Promise<Organization> {
-        return await $api.put(`/api/v1/organizations/${orgId}/approve`)
+        let response = await $api.put<Organization>(`/api/v1/admin/organizations/${orgId}/approve`);
+        return response.data;
     }
 
     async delete(orgId: number): Promise<null> {
-        return await $api.delete(`/api/v1/organizations/${orgId}`)
+        return await $api.delete(`/api/v1/admin/organizations/${orgId}`)
     }
 }
